modules/api_grpc/requirements.txt
grpcio-tools
kafka-python